refactor: QUZ-43 패키지 구조 변경
