+++ user-service/user-domain/src/main/kotlin/com/grepp/quizy/domain/user/AuthProvider.kt
@@ -1,4 +1,4 @@
-package com.grepp.quizy.user
+package com.grepp.quizy.domain.user
 
 enum class AuthProvider {
     GOOGLE,
