+++ user-service/user-infra/src/main/kotlin/com/grepp/quizy/infra/user/exception/UserNotFoundException.kt
@@ -1,6 +1,7 @@
-package com.grepp.quizy.user.exception
+package com.grepp.quizy.infra.user.exception
 
 import com.grepp.quizy.common.exception.InfraException
+import com.grepp.quizy.domain.user.exception.UserErrorCode
 
 class UserNotFoundException : InfraException(UserErrorCode.USER_NOT_FOUND) {
